Refactor some of the core APIs to consume and operate on ResolverContext
A ResolverContext acts as an intermediate entity that encapsulates all the information that a resolver would potentially need but can be constructed before the resolver has been identified

So far we have assumed that all operations are coupled to a single store and single resolver. But this is inadequate in many scenarios.
This refactoring paves the way for supporting multi-resolver resolver operations, multi-source operations (eg. operations like union queries which merge in information from multiple sorces) and operations having no coupled sources (eg. operations that just delegate to one or more of other resolvers, and so are not directly associated with a particular source).